move from draw-map.js to added synteny-blocks-draw.js : syntenyBlock_2Feature,  showSynteny().
patham2() will be moved in next commit to make it accessible to showSynteny().
